Struct challonge::attachments::AttachmentCreate

pub struct AttachmentCreate {
    pub asset: Option<Vec<u8>>,
    pub url: Option<String>,
    pub description: Option<String>,
}

A structure for creating an attachment * At least 1 of the 3 optional parameters must be provided. * Files up to 25MB are allowed for tournaments hosted by Premier badge Challonge Premier subscribers.

Fields

asset

A file upload (250KB max, no more than 4 attachments per match). If provided, the url parameter will be ignored.

url

A web (http, ftp) link

description

Text to describe the file or URL attachment, or this can simply be standalone text.

Methods

impl AttachmentCreate

fn new() -> AttachmentCreate

Creates new AttachmentCreate structure with default values.

fn asset<'a>(&'a mut self, asset: Vec<u8>) -> &'a mut Self

A builder method for $field with Option type.

fn url<'a, S: Into<String>>(&'a mut self, url: S) -> &'a mut Self

A builder method for $field with Option<String> type.

fn description<'a, S: Into<String>>(&'a mut self, description: S) -> &'a mut Self

A builder method for $field with Option<String> type.