Struct challonge::matches::MatchUpdate

pub struct MatchUpdate {
    pub scores_csv: MatchScores,
    pub winner_id: Option<ParticipantId>,
    pub player1_votes: Option<u64>,
    pub player2_votes: Option<u64>,
}

NOTE: If you're updating winner_id, scores_csv must also be provided. You may, however, update score_csv without providing winner_id for live score updates.

Fields

scores_csv

Comma separated set/game scores with player 1 score first (e.g. "1-3,3-0,3-2")

winner_id

The participant ID of the winner or "tie" if applicable (Round Robin and Swiss). NOTE: If you change the outcome of a completed match, all matches in the bracket that branch from the updated match will be reset.

player1_votes

Overwrites the number of votes for player 1

player2_votes

Overwrites the number of votes for player 2

Methods

impl MatchUpdate

fn new() -> MatchUpdate

Creates new MatchUpdate structure with default values.

fn scores_csv<'a>(&'a mut self, scores_csv: MatchScores) -> &'a mut Self

A builder method for $field with $field_type type.

fn winner_id<'a>(&'a mut self, winner_id: ParticipantId) -> &'a mut Self

A builder method for $field with Option type.

fn player1_votes<'a>(&'a mut self, player1_votes: u64) -> &'a mut Self

A builder method for $field with Option type.

fn player2_votes<'a>(&'a mut self, player2_votes: u64) -> &'a mut Self

A builder method for $field with Option type.