Struct challonge::Challonge

pub struct Challonge {
    // some fields omitted
}

Client for the Challonge REST API.

Methods

impl Challonge

fn new(user_name: &str, api_key: &str) -> Challonge

Create new connection to Challonge.

Example

extern crate challonge;

use self::challonge::Challonge;

let c = Challonge::new("myusername", "myapikey");

fn tournament_index(&self, state: &TournamentState, tournament_type: &TournamentType, created_after: &Date<Local>, created_before: &Date<Local>, subdomain: &str) -> Result<TournamentIndex, Error>

Retrieve a set of tournaments created with your account.

Example

extern crate challonge;
extern crate chrono;

use self::challonge::Challonge;
use self::challonge::tournament::{ TournamentState, TournamentType };
use self::chrono::*;

let c = Challonge::new("myusername", "myapikey");
let index = c.tournament_index (
       &TournamentState::All,
       &TournamentType::DoubleElimination,
       &Local::today(),
       &Local::today(),
       "subdomain"
);

fn get_tournament(&self, id: &TournamentId, include_participants: bool, include_matches: bool) -> Result<Tournament, Error>

Retrieve a single tournament record created with your account.

Example

extern crate challonge;

use challonge::Challonge;

let c = Challonge::new("myusername", "myapikey");
let t = c.get_tournament(&TournamentId::Id(2669881), true, true);

fn create_tournament(&self, tournament: TournamentCreate) -> Result<Tournament, Error>

Create a new tournament.

Example

extern crate challonge;

use challonge::Challonge;
use challonge::tournament::TournamentCreate;

let c = Challonge::new("myusername", "myapikey");
let tc = TournamentCreate {
           name: "Tester".to_owned(),
           tournament_type: TournamentType::SingleElimination,
           url: "testerurl".to_owned(),
           subdomain: "subdomain".to_owned(),
           description: "Test tournament created from challonge-rs".to_owned(),
           open_signup: false,
           hold_third_place_match: false,
           pts_for_match_win: 0.0f64,
           pts_for_match_tie: 0.0f64,
           pts_for_game_win: 0.0f64,
           pts_for_game_tie: 0.0f64,
           pts_for_bye: 0.0f64,
           swiss_rounds: 0,
           ranked_by: RankedBy::PointsScored,
           rr_pts_for_match_win: 0.0f64,
           rr_pts_for_match_tie: 0.0f64,
           rr_pts_for_game_win: 0.0f64,
           rr_pts_for_game_tie: 0.0f64,
           show_rounds: false,
           private: false,
           notify_users_when_matches_open: true,
           notify_users_when_the_tournament_ends: true,
           sequential_pairings: false,
           signup_cap: 4,
           start_at: UTC::now().add(Duration::weeks(2)),
           check_in_duration: 60,
           grand_finals_modifier: None,
};
let t = c.create_tournament(tc.clone());
assert!(t.is_ok());

fn update_tournament(&self, id: TournamentId, tournament: TournamentCreate) -> Result<Tournament, Error>

Update a tournament's attributes.

fn delete_tournament(&self, id: TournamentId) -> Result<(), Error>

Deletes a tournament along with all its associated records. There is no undo, so use with care!

fn tournament_process_checkins(&self, id: TournamentId, include_participants: bool, include_matches: bool) -> Result<(), Error>

This should be invoked after a tournament's check-in window closes before the tournament is started.

Marks participants who have not checked in as inactive.

Moves inactive participants to bottom seeds (ordered by original seed).

Transitions the tournament state from 'checking_in' to 'checked_in'

NOTE: Checked in participants on the waiting list will be promoted if slots become available.

fn tournament_abort_checkins(&self, id: TournamentId, include_participants: bool, include_matches: bool) -> Result<(), Error>

When your tournament is in a 'checking_in' or 'checked_in' state, there's no way to edit the tournament's start time (start_at) or check-in duration (check_in_duration). You must first abort check-in, then you may edit those attributes.

Makes all participants active and clears their checked_in_at times.

Transitions the tournament state from 'checking_in' or 'checked_in' to 'pending'

fn tournament_start(&self, id: TournamentId, include_participants: bool, include_matches: bool) -> Result<(), Error>

Start a tournament, opening up first round matches for score reporting. The tournament must have at least 2 participants.

fn tournament_finalize(&self, id: TournamentId, include_participants: bool, include_matches: bool) -> Result<(), Error>

Finalize a tournament that has had all match scores submitted, rendering its results permanent.

fn tournament_reset(&self, id: TournamentId, include_participants: bool, include_matches: bool) -> Result<(), Error>

Reset a tournament, clearing all of its scores and attachments. You can then add/remove/edit participants before starting the tournament again.

fn participant_index(&self, id: TournamentId) -> Result<ParticipantIndex, Error>

Retrieve a tournament's participant list.

fn create_participant(&self, id: TournamentId, participant: ParticipantCreate) -> Result<Participant, Error>

Add a participant to a tournament (up until it is started).

fn get_participant(&self, id: TournamentId, participant_id: ParticipantId, include_matches: bool) -> Result<Participant, Error>

Retrieve a single participant record for a tournament.

fn update_participant(&self, id: TournamentId, participant_id: ParticipantId, participant: ParticipantCreate) -> Result<(), Error>

Update the attributes of a tournament participant.

fn check_in_participant(&self, id: TournamentId, participant_id: ParticipantId) -> Result<(), Error>

Checks a participant in, setting checked_in_at to the current time.

fn undo_check_in_participant(&self, id: TournamentId, participant_id: ParticipantId) -> Result<(), Error>

Marks a participant as having not checked in, setting checked_in_at to nil.

fn delete_participant(&self, id: TournamentId, participant_id: ParticipantId) -> Result<(), Error>

If the tournament has not started, delete a participant, automatically filling in the abandoned seed number. If tournament is underway, mark a participant inactive, automatically forfeiting his/her remaining matches.

fn randomize_participants(&self, id: TournamentId) -> Result<(), Error>

Randomize seeds among participants. Only applicable before a tournament has started.

fn match_index(&self, id: TournamentId, state: Option<MatchState>, participant_id: Option<ParticipantId>) -> Result<MatchIndex, Error>

Retrieve a tournament's match list.

fn get_match(&self, id: TournamentId, match_id: MatchId, include_attachments: bool) -> Result<Match, Error>

Retrieve a single match record for a tournament.

fn update_match(&self, id: TournamentId, match_id: MatchId, match_update: MatchUpdate) -> Result<Match, Error>

Update/submit the score(s) for a match.