Struct challonge::tournament::TournamentCreate

pub struct TournamentCreate {
    pub name: String,
    pub tournament_type: TournamentType,
    pub url: String,
    pub subdomain: String,
    pub description: String,
    pub open_signup: bool,
    pub hold_third_place_match: bool,
    pub swiss_points: GamePoints,
    pub swiss_rounds: u64,
    pub ranked_by: RankedBy,
    pub round_robin_points: GamePoints,
    pub show_rounds: bool,
    pub private: bool,
    pub notify_users_when_matches_open: bool,
    pub notify_users_when_the_tournament_ends: bool,
    pub sequential_pairings: bool,
    pub signup_cap: u64,
    pub start_at: Option<DateTime<UTC>>,
    pub check_in_duration: u64,
    pub grand_finals_modifier: Option<String>,
}

Structure for creating a tournament.

Fields

name

Your event's name/title (Max: 60 characters)

tournament_type

Type of a tournament

url

challonge.com/url (letters, numbers, and underscores only)

subdomain

subdomain.challonge.com/url (Requires write access to the specified subdomain)

description

Description/instructions to be displayed above the bracket

open_signup

Have Challonge host a sign-up page (otherwise, you manually add all participants)

hold_third_place_match

Single Elimination only

swiss_points

Only for Swiss system

swiss_rounds

Number of rounds in swiss system

ranked_by

Tournament ranking type

round_robin_points

Only for Round Robin system

show_rounds

Single & Double Elimination only - Label each round above the bracket (default: false)

private

Hide this tournament from the public browsable index and your profile (default: false)

notify_users_when_matches_open

Email registered Challonge participants when matches open up for them (default: false)

notify_users_when_the_tournament_ends

Email registered Challonge participants the results when this tournament ends (default: false)

sequential_pairings

Instead of traditional seeding rules, make pairings by going straight down the list of participants. First round matches are filled in top to bottom, then qualifying matches (if applicable). (default: false)

signup_cap

Maximum number of participants in the bracket. A waiting list (attribute on Participant) will capture participants once the cap is reached.

start_at

the planned or anticipated start time for the tournament (Used with check_in_duration to determine participant check-in window). Timezone defaults to Eastern.

check_in_duration

Length of the participant check-in window in minutes.

grand_finals_modifier

This option only affects double elimination. null/blank (default) - give the winners bracket finalist two chances to beat the losers bracket finalist, 'single match' - create only one grand finals match, 'skip' - don't create a finals match between winners and losers bracket finalists

Methods

impl TournamentCreate

fn new() -> TournamentCreate

Creates new TournamentCreate structure with default values.

fn name<'a, S: Into<String>>(&'a mut self, name: S) -> &'a mut Self

A builder method for $field with String type.

fn tournament_type<'a>(&'a mut self, tournament_type: TournamentType) -> &'a mut Self

A builder method for $field with $field_type type.

fn url<'a, S: Into<String>>(&'a mut self, url: S) -> &'a mut Self

A builder method for $field with String type.

fn subdomain<'a, S: Into<String>>(&'a mut self, subdomain: S) -> &'a mut Self

A builder method for $field with String type.

fn description<'a, S: Into<String>>(&'a mut self, description: S) -> &'a mut Self

A builder method for $field with String type.

fn open_signup<'a>(&'a mut self, open_signup: bool) -> &'a mut Self

A builder method for $field with $field_type type.

fn hold_third_place_match<'a>(&'a mut self, hold_third_place_match: bool) -> &'a mut Self

A builder method for $field with $field_type type.

fn swiss_points<'a>(&'a mut self, swiss_points: GamePoints) -> &'a mut Self

A builder method for $field with $field_type type.

fn swiss_rounds<'a>(&'a mut self, swiss_rounds: u64) -> &'a mut Self

A builder method for $field with $field_type type.

fn ranked_by<'a>(&'a mut self, ranked_by: RankedBy) -> &'a mut Self

A builder method for $field with $field_type type.

fn round_robin_points<'a>(&'a mut self, round_robin_points: GamePoints) -> &'a mut Self

A builder method for $field with $field_type type.

fn show_rounds<'a>(&'a mut self, show_rounds: bool) -> &'a mut Self

A builder method for $field with $field_type type.

fn private<'a>(&'a mut self, private: bool) -> &'a mut Self

A builder method for $field with $field_type type.

fn notify_users_when_matches_open<'a>(&'a mut self, notify_users_when_matches_open: bool) -> &'a mut Self

A builder method for $field with $field_type type.

fn notify_users_when_the_tournament_ends<'a>(&'a mut self, notify_users_when_the_tournament_ends: bool) -> &'a mut Self

A builder method for $field with $field_type type.

fn sequential_pairings<'a>(&'a mut self, sequential_pairings: bool) -> &'a mut Self

A builder method for $field with $field_type type.

fn signup_cap<'a>(&'a mut self, signup_cap: u64) -> &'a mut Self

A builder method for $field with $field_type type.

fn check_in_duration<'a>(&'a mut self, check_in_duration: u64) -> &'a mut Self

A builder method for $field with $field_type type.

fn grand_finals_modifier<'a>(&'a mut self, grand_finals_modifier: Option<String>) -> &'a mut Self

A builder method for $field with $field_type type.