Struct challonge::tournament::TournamentCreate

pub struct TournamentCreate {
    pub name: String,
    pub tournament_type: TournamentType,
    pub url: String,
    pub subdomain: String,
    pub description: String,
    pub open_signup: bool,
    pub hold_third_place_match: bool,
    pub pts_for_match_win: f64,
    pub pts_for_match_tie: f64,
    pub pts_for_game_win: f64,
    pub pts_for_game_tie: f64,
    pub pts_for_bye: f64,
    pub swiss_rounds: u64,
    pub ranked_by: RankedBy,
    pub rr_pts_for_match_win: f64,
    pub rr_pts_for_match_tie: f64,
    pub rr_pts_for_game_win: f64,
    pub rr_pts_for_game_tie: f64,
    pub show_rounds: bool,
    pub private: bool,
    pub notify_users_when_matches_open: bool,
    pub notify_users_when_the_tournament_ends: bool,
    pub sequential_pairings: bool,
    pub signup_cap: u64,
    pub start_at: DateTime<UTC>,
    pub check_in_duration: u64,
    pub grand_finals_modifier: Option<String>,
}

Structure for creating a tournament.

Fields

name

Your event's name/title (Max: 60 characters)

tournament_type

url

challonge.com/url (letters, numbers, and underscores only)

subdomain

subdomain.challonge.com/url (Requires write access to the specified subdomain)

description

Description/instructions to be displayed above the bracket

open_signup

Have Challonge host a sign-up page (otherwise, you manually add all participants)

hold_third_place_match

Single Elimination only

pts_for_match_win

Only for Swiss system

pts_for_match_tie

pts_for_game_win

pts_for_game_tie

pts_for_bye

swiss_rounds

ranked_by

rr_pts_for_match_win

Only for Round Robin system

rr_pts_for_match_tie

rr_pts_for_game_win

rr_pts_for_game_tie

show_rounds

Single & Double Elimination only - Label each round above the bracket (default: false)

private

Hide this tournament from the public browsable index and your profile (default: false)

notify_users_when_matches_open

Email registered Challonge participants when matches open up for them (default: false)

notify_users_when_the_tournament_ends

Email registered Challonge participants the results when this tournament ends (default: false)

sequential_pairings

Instead of traditional seeding rules, make pairings by going straight down the list of participants. First round matches are filled in top to bottom, then qualifying matches (if applicable). (default: false)

signup_cap

Maximum number of participants in the bracket. A waiting list (attribute on Participant) will capture participants once the cap is reached.

start_at

the planned or anticipated start time for the tournament (Used with check_in_duration to determine participant check-in window). Timezone defaults to Eastern.

check_in_duration

Length of the participant check-in window in minutes.

grand_finals_modifier

This option only affects double elimination. null/blank (default) - give the winners bracket finalist two chances to beat the losers bracket finalist, 'single match' - create only one grand finals match, 'skip' - don't create a finals match between winners and losers bracket finalists